[Diffusion] Unify LTX forward request handling
