+++ vllm_omni/diffusion/models/ltx2/ltx2_conditioning.py
@@ -297,11 +297,6 @@ class LTXI2VConditioningMixin:
             resample="bilinear",
         )
 
-    @torch.no_grad()
-    def forward(self, req: Any, image: Any | None = None, *args: Any, **kwargs: Any) -> Any:
-        """Route I2V inputs through the task-neutral private runner."""
-        return self._forward_request(req, *args, image=image, **kwargs)
-
     @staticmethod
     def _resolve_single_prompt_image(raw_image: Any) -> Any:
         if isinstance(raw_image, list):
